+++ src/main/java/com/fasterxml/jackson/databind/util/clhm/ConcurrentLinkedHashMap.java
@@ -32,7 +32,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fasterxml.jackson.databind.util.concurrentlinkedhashmap;
+package com.fasterxml.jackson.databind.util.clhm;
 
 import java.io.InvalidObjectException;
 import java.io.ObjectInputStream;
