+++ src/main/java/com/fasterxml/jackson/databind/util/clhm/EvictionListener.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fasterxml.jackson.databind.util.concurrentlinkedhashmap;
+package com.fasterxml.jackson.databind.util.clhm;
 
 /**
  * A listener registered for notification when an entry is evicted. An instance
