+++ src/test/java/com/fasterxml/jackson/databind/util/clhm/ConcurrentLinkedHashMapTest.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  ****************************************************************/
-package com.fasterxml.jackson.databind.util.concurrentlinkedhashmap;
+package com.fasterxml.jackson.databind.util.clhm;
 
 import org.junit.Test;
